# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Digital Infrastructure for Sri Lankan Schools",
        "title": "OpenSchool API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/enrollments/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Dry run: checks picks against every group of the level without saving",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Validate enrollment picks",
                "parameters": [
                    {
                        "description": "Proposed picks",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubmitEnrollmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EnrollmentValidationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/groups/{group_id}/students": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "List students enrolled through a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group UUID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Academic year UUID",
                        "name": "academic_year_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EnrolledStudentResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/students/{id}/enrollments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "List a student's subjects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Academic year UUID",
                        "name": "academic_year_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EnrollmentResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates then replaces the student's picks for that level and academic year",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Submit enrollment picks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Picks",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SubmitEnrollmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EnrollmentValidationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.EnrollmentValidationResponse"
                        }
                    }
                }
            }
        },
        "/students/{id}/enrollments/lock/{level_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only: removes the lock so the student's picks for this level+year can be changed again",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Unlock a student's confirmed subject selection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Level UUID",
                        "name": "level_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Academic year UUID",
                        "name": "academic_year_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/students/{id}/enrollments/{group_id}/{subject_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "Remove one enrollment pick",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group UUID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subject UUID",
                        "name": "subject_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Academic year UUID",
                        "name": "academic_year_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/subjects/{id}/students": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enrollments"
                ],
                "summary": "List students taking a subject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subject UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Academic year UUID",
                        "name": "academic_year_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EnrolledStudentResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.EnrolledStudentResponse": {
            "type": "object",
            "properties": {
                "enrolled_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "group_label": {
                    "type": "string"
                },
                "index_number": {
                    "type": "string"
                },
                "medium_id": {
                    "type": "string"
                },
                "medium_name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "subject_code": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "subject_name": {
                    "type": "string"
                }
            }
        },
        "models.EnrollmentPick": {
            "type": "object",
            "required": [
                "group_id",
                "subject_id"
            ],
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "medium_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "models.EnrollmentResponse": {
            "type": "object",
            "properties": {
                "academic_year_id": {
                    "type": "string"
                },
                "enrolled_at": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "group_label": {
                    "type": "string"
                },
                "level_id": {
                    "type": "string"
                },
                "level_label": {
                    "type": "string"
                },
                "medium_id": {
                    "type": "string"
                },
                "medium_name": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "subject_code": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "subject_name": {
                    "type": "string"
                },
                "subject_type": {
                    "type": "string"
                }
            }
        },
        "models.EnrollmentValidationResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GroupValidationError"
                    }
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "models.GroupValidationError": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.SubmitEnrollmentRequest": {
            "type": "object",
            "required": [
                "academic_year_id",
                "level_id"
            ],
            "properties": {
                "academic_year_id": {
                    "type": "string"
                },
                "level_id": {
                    "type": "string"
                },
                "picks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EnrollmentPick"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT token. Example: \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}